+++ CHANGELOG.md
@@ -4,5 +4,5 @@
 - Change 2020V1, 2020V2 LVGL touch I2C polling to interrupt polling, others are not currently adapted
 - Added TTGOClass mutex example under multitasking <examples\BasicUnit\TouchMultitask>
 - Added setting touch default monitoring report rate in touch initialization and enter monitoring mode in active state Description
-
+- Added 2020V2 touch wakeup method, touch can be set to sleep in 2020V2, control AXP202_EXTEN can wake up touch